V_9
the best model till now!
So we are still using our LightGBM and this time we have fully optimised the threshold and the accuracies are very good. Also to make the thresholds better, instead of using the accuracy i have used the F1-scores to ensure that there is a balance and not just the accuracy.

New streamlit design, so added better and easier to follow layout, next i would want to work on the local vs global SHAP features so we can see exactly how each feature affected the model's predictions.