new update
oh lord, so i did so much for this update. I've tried using 4 different XAI methods and then compare them but we always had problems with the attention and sometimes the results couldn't be trusted. so we will be moving to version 5 and aim to solve this "Conformal Prediction (CP) for multi-label toxicity", we are still using the based model from v 4 but this time with new changes.